api/requirements.txt
Flask
psycopg2-binary
requests
Flask-cors